Prepare v2016.01

Various QSPI fixes from Siva were applied:
spi_flash: Add support for device with Quad IO support
zynq_qspi: Update parent priv data during pre probe of qspi
zynq-common: Add missing dual flash support in zynq
spi: spi_flash: Add support for generic qspi
zynqmp_qspi: Update parent priv data during pre probe of qspi
spi: spi_flash: Add 4-byte support for spi flash
spi: spi_flash: Correct flash size calculation
spi: spi_flash: Add support for Dual stacked mode

Add ifdef around CONFIG_FEC_MXC to remove compilation warnings.
Remove qspi flash enabling via defconfig. Do it later.

+++ arch/arm/include/asm/arch-hi6220/gpio.h
@@ -21,7 +21,7 @@ struct gpio_bank {
 /* Information about a GPIO bank */
 struct hikey_gpio_platdata {
 	int bank_index;
-	unsigned int base;     /* address of registers in physical memory */
+	ulong base;     /* address of registers in physical memory */
 };
 
 #endif /* _HI6220_GPIO_H_ */
