+++ drivers/usb/dwc3/io.h
@@ -31,8 +31,8 @@ static inline u32 dwc3_readl(void __iomem *base, u32 offset)
 	 * space, see dwc3_probe in core.c.
 	 * However, the offsets are given starting from xHCI address space.
 	 */
-	offs += base;
-	value = readl(offs);
+	offs += (unsigned long)base;
+	value = readl((unsigned long)offs);
 
 	return value;
 }
@@ -46,8 +46,8 @@ static inline void dwc3_writel(void __iomem *base, u32 offset, u32 value)
 	 * space, see dwc3_probe in core.c.
 	 * However, the offsets are given starting from xHCI address space.
 	 */
-	offs += base;
-	writel(value, offs);
+	offs += (unsigned long)base;
+	writel(value, (unsigned long)offs);
 }
 
 static inline void dwc3_flush_cache(int addr, int length)
